requirements.txt
Django==2.2.6
psycopg2==2.8.3
djangorestframework==3.10.3
markdown==3.1.1
django-filter==2.2.0
gunicorn==19.7.1
pytz==2017.3
whitenoise==3.3.1
dj-database-url==0.4.2
django-cors-headers
